backend/AI_chat_companion/requirements.txt
fastapi
uvicorn
langchain
chromadb
huggingface-hub
google-generativeai
pydantic
python-multipart
requests
transformers